+++ examples/flutter_integration_tests/integration_test/integration_test.dart
@@ -110,5 +110,5 @@ void main() {
 }
 
 extension on Database {
-  void closeWhenDone() => addTearDown(dispose);
+  void closeWhenDone() => addTearDown(close);
 }