+++ sqlite3/example/custom_extension/test/extension_test.dart
@@ -9,7 +9,7 @@ void main() {
 
   test('can use vec0', () {
     final db = sqlite3.openInMemory();
-    addTearDown(db.dispose);
+    addTearDown(db.close);
 
     db.execute(
       'create virtual table vec_examples using vec0(sample_embedding float[8])',