+++ sqlite3/example/web/worker.dart
@@ -39,7 +39,7 @@ void main() {
         ..execute('pragma user_version = 1')
         ..execute('CREATE TABLE foo (bar INTEGER NOT NULL);')
         ..execute('INSERT INTO foo (bar) VALUES (?)', [3])
-        ..dispose();
+        ..close();
 
       final db = sqlite3.open('/database');
       print(db.select('SELECT * FROM foo'));