+++ sqlite3/test/common/prepared_statement.dart
@@ -19,7 +19,7 @@ void testPreparedStatements(
     final db = sqlite3.openInMemory()
       ..execute('CREATE TABLE foo (a INTEGER);')
       ..execute('CREATE TABLE télé (a INTEGER);');
-    addTearDown(db.dispose);
+    addTearDown(db.close);
 
     final stmts = db.prepareMultiple('SELECT * FROM foo;SELECT * FROM télé;');
 
@@ -34,7 +34,7 @@ void testPreparedStatements(
     final stmt = opened.prepare('INSERT INTO tbl(a) VALUES(?)');
     stmt.execute(<String>['a']);
     stmt.execute(<String>['b']);
-    stmt.dispose();
+    stmt.close();
 
     final select = opened.prepare('SELECT * FROM tbl ORDER BY a');
     final result = select.select();
@@ -42,14 +42,14 @@ void testPreparedStatements(
     expect(result, hasLength(2));
     expect(result.map((row) => row['a'] as String), ['a', 'b']);
 
-    select.dispose();
+    select.close();
 
-    opened.dispose();
+    opened.close();
   });
 
   test('prepared statements without parameters can be used multiple times', () {
     final opened = sqlite3.openInMemory();
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
     opened
       ..execute('CREATE TABLE tbl (a TEXT);')
       ..execute('INSERT INTO tbl DEFAULT VALUES;');
@@ -63,23 +63,23 @@ void testPreparedStatements(
     final opened = sqlite3.openInMemory();
 
     final stmt = opened.prepare('SELECT ?');
-    stmt.dispose();
+    stmt.close();
 
     expect(stmt.select, throwsStateError);
-    opened.dispose();
+    opened.close();
   });
 
   test('prepared statements cannot be used after db is closed', () {
     final opened = sqlite3.openInMemory();
     final stmt = opened.prepare('SELECT 1');
-    opened.dispose();
+    opened.close();
 
     expect(stmt.select, throwsStateError);
   });
 
   test('parameterCount', () {
     final opened = sqlite3.openInMemory();
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
 
     expect(opened.prepare('SELECT 1').parameterCount, 0);
     expect(opened.prepare('SELECT 1, ?2 AS r').parameterCount, 2);
@@ -88,7 +88,7 @@ void testPreparedStatements(
   test('isReadOnly', () {
     final opened = sqlite3.openInMemory()
       ..execute('CREATE TABLE tbl (a TEXT);');
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
 
     expect(opened.prepare('SELECT 1').isReadOnly, isTrue);
     expect(opened.prepare('UPDATE tbl SET a = a || ?').isReadOnly, isFalse);
@@ -97,7 +97,7 @@ void testPreparedStatements(
   test('isExplain', () {
     final opened = sqlite3.openInMemory()
       ..execute('CREATE TABLE tbl (a TEXT);');
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
 
     expect(opened.prepare('SELECT 1').isExplain, isFalse);
     expect(opened.prepare('EXPLAIN SELECT 1').isExplain, isTrue);
@@ -109,12 +109,12 @@ void testPreparedStatements(
 
     final insert = opened.prepare('INSERT INTO tbl VALUES (?)');
     insert.execute(<dynamic>[value]);
-    insert.dispose();
+    insert.close();
 
     final select = opened.prepare('SELECT * FROM tbl');
     final result = select.select().single;
 
-    opened.dispose();
+    opened.close();
     return result['x'] as Uint8List?;
   }
 
@@ -148,7 +148,7 @@ void testPreparedStatements(
       ),
     );
 
-    db.dispose();
+    db.close();
   });
 
   test('throws an exception when iterating over result rows', () {
@@ -193,7 +193,7 @@ void testPreparedStatements(
     final result = stmt.select([false]).single;
 
     expect(result.values.single, isZero);
-    db.dispose();
+    db.close();
   });
 
   test('can bind named parameters', () {
@@ -214,7 +214,7 @@ void testPreparedStatements(
 
   test('can bind custom values', () {
     final db = sqlite3.openInMemory();
-    addTearDown(db.dispose);
+    addTearDown(db.close);
 
     final stmt = db.prepare('SELECT :a AS a, :b AS b');
     final result = stmt.selectWith(
@@ -233,7 +233,7 @@ void testPreparedStatements(
     late CommonDatabase db;
 
     setUp(() => db = sqlite3.openInMemory());
-    tearDown(() => db.dispose());
+    tearDown(() => db.close());
 
     test('when no parameters are set', () {
       final stmt = db.prepare('SELECT ?');
@@ -289,12 +289,12 @@ void testPreparedStatements(
     expect(result3.columnNames, ['?']);
     expect(result3.single.columnAt(0), '');
 
-    opened.dispose();
+    opened.close();
   });
 
   test('does not validate custom parameters', () {
     final opened = sqlite3.openInMemory();
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
 
     final stmt = opened.prepare('SELECT ? AS r');
     expect(stmt.selectWith(StatementParameters.bindCustom((stmt) {})), [
@@ -306,7 +306,7 @@ void testPreparedStatements(
     final opened = sqlite3.openInMemory()
       ..execute('create table t (c1)')
       ..execute('insert into t values (1)');
-    addTearDown(opened.dispose);
+    addTearDown(opened.close);
 
     final stmt = opened.prepare('select * from t');
     expect(stmt.select(), [
@@ -338,7 +338,7 @@ void testPreparedStatements(
       expect(cursor.moveNext(), isFalse);
 
       opened.execute('commit');
-      opened.dispose();
+      opened.close();
     },
     skip: supportsReturning
         ? null
@@ -350,7 +350,7 @@ void testPreparedStatements(
 
     setUp(() => database = sqlite3.openInMemory());
 
-    tearDown(() => database.dispose());
+    tearDown(() => database.close());
 
     test('report correct values', () {
       final stmt = database.prepare('VALUES (1), (2), (3);');
@@ -407,7 +407,7 @@ void testPreparedStatements(
       final opened = sqlite3.openInMemory()
         ..execute('create table t (c1)')
         ..execute('insert into t values (1)');
-      addTearDown(opened.dispose);
+      addTearDown(opened.close);
 
       final stmt = opened.prepare('select * from t');
       var cursor = stmt.selectCursor();
@@ -431,7 +431,7 @@ void testPreparedStatements(
         ..execute('create table t (c1)')
         ..execute('insert into t values (1)')
         ..execute('insert into t values (2)');
-      addTearDown(opened.dispose);
+      addTearDown(opened.close);
 
       final stmt = opened.prepare('select * from t');
       final cursor = stmt.selectCursor();
@@ -453,7 +453,7 @@ void testPreparedStatements(
         final cursor = stmt.selectCursor();
         expect(cursor.moveNext(), isTrue);
 
-        stmt.dispose();
+        stmt.close();
         expect(cursor.moveNext(), isFalse);
       });
 
@@ -464,7 +464,7 @@ void testPreparedStatements(
 
         stmt.reset();
         expect(cursor.moveNext(), isFalse);
-        stmt.dispose();
+        stmt.close();
       });
 
       test('by invoking select', () {
@@ -511,8 +511,8 @@ void testPreparedStatements(
       });
 
       tearDown(() {
-        statement.dispose();
-        database.dispose();
+        statement.close();
+        database.close();
       });
 
       test('can be used with execute', () {
@@ -536,7 +536,7 @@ void testPreparedStatements(
     late CommonDatabase db;
 
     setUp(() => db = sqlite3.openInMemory());
-    tearDown(() => db.dispose());
+    tearDown(() => db.close());
 
     test('for syntax', () {
       final throwsSyntaxError = throwsSqlError(1, 1);
@@ -597,12 +597,12 @@ void testPreparedStatements(
       addTearDown(() => sqlite3.unregisterVirtualFileSystem(vfs));
 
       var db = sqlite3.open('/db', vfs: vfs.name);
-      addTearDown(() => db.dispose());
+      addTearDown(() => db.close());
 
       db
         ..execute('CREATE TABLE foo (bar TEXT) STRICT')
         ..execute('INSERT INTO foo (bar) VALUES (?)', ['testing'])
-        ..dispose();
+        ..close();
 
       db = db = sqlite3.open('/db', vfs: vfs.name);
       final stmt = db.prepare('SELECT * FROM foo');
@@ -617,7 +617,7 @@ void testPreparedStatements(
       expect(cursor.moveNext(), isTrue);
       expect(cursor.current, {'bar': 'testing'});
 
-      stmt.dispose();
+      stmt.close();
     });
   });
 }